requirements/base.txt
psycopg2-binary==2.8.3
gunicorn==19.9.0

Django==2.2.6
djangorestframework==3.10.3 

requests==2.22.0
coverage==4.5.4
